// src/types/glTF.d.ts
export interface GlTF {
  /**
   * An array of accessors.
   */
  accessors?: Accessor[]
  /**
   * An array of keyframe animations.
   */
  animations?: Animation[]
  /**
   * Metadata about the glTF asset.
   */
  asset: Asset
  /**
   * An array of buffers.
   */
  buffers?: Buffer[]
  /**
   * An array of bufferViews.
   */
  bufferViews?: BufferView[]
  /**
   * An array of cameras.
   */
  cameras?: Camera[]
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Names of glTF extensions required to properly load this asset.
   */
  extensionsRequired?: string[]
  /**
   * Names of glTF extensions used somewhere in this asset.
   */
  extensionsUsed?: string[]
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * An array of images.
   */
  images?: Image[]
  /**
   * An array of materials.
   */
  materials?: Material[]
  /**
   * An array of meshes.
   */
  meshes?: Mesh[]
  /**
   * An array of nodes.
   */
  nodes?: Node[]
  /**
   * An array of samplers.
   */
  samplers?: Sampler[]
  /**
   * The index of the default scene.
   */
  scene?: number
  /**
   * An array of scenes.
   */
  scenes?: Scene[]
  /**
   * An array of skins.
   */
  skins?: Skin[]
  /**
   * An array of textures.
   */
  textures?: Texture[]
}

/**
 * A typed view into a bufferView.  A bufferView contains raw binary data.  An accessor
 * provides a typed view into a bufferView or a subset of a bufferView similar to how
 * WebGL's `vertexAttribPointer()` defines an attribute in a buffer.
 */
export interface Accessor {
  /**
   * The index of the bufferView.
   */
  bufferView?: number
  /**
   * The offset relative to the start of the bufferView in bytes.
   */
  byteOffset?: number
  /**
   * The datatype of components in the attribute.
   */
  componentType: number
  /**
   * The number of attributes referenced by this accessor.
   */
  count: number
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * Maximum value of each component in this attribute.
   */
  max?: number[]
  /**
   * Minimum value of each component in this attribute.
   */
  min?: number[]
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * Specifies whether integer data values should be normalized.
   */
  normalized?: boolean
  /**
   * Sparse storage of attributes that deviate from their initialization value.
   */
  sparse?: AccessorSparse
  /**
   * Specifies if the attribute is a scalar, vector, or matrix.
   */
  type: string
}

/**
 * Sparse storage of attributes that deviate from their initialization value.
 */
export interface AccessorSparse {
  /**
   * Number of entries stored in the sparse array.
   */
  count: number
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * Index array of size `count` that points to those accessor attributes that deviate from
   * their initialization value. Indices must strictly increase.
   */
  indices: AccessorSparseIndices
  /**
   * Array of size `count` times number of components, storing the displaced accessor
   * attributes pointed by `indices`. Substituted values must have the same `componentType`
   * and number of components as the base accessor.
   */
  values: AccessorSparseValues
}
// ...
export interface Animation {
  /**
   * An array of channels, each of which targets an animation's sampler at a node's property.
   * Different channels of the same animation can't have equal targets.
   */
  channels: AnimationChannel[]
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * An array of samplers that combines input and output accessors with an interpolation
   * algorithm to define a keyframe graph (but not its target).
   */
  samplers: AnimationSampler[]
}
// ...
export interface Asset {
  /**
   * A copyright message suitable for display to credit the content creator.
   */
  copyright?: string
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * Tool that generated this glTF model.  Useful for debugging.
   */
  generator?: string
  /**
   * The minimum glTF version that this asset targets.
   */
  minVersion?: string
  /**
   * The glTF version that this asset targets.
   */
  version: string
}

/**
 * A view into a buffer generally representing a subset of the buffer.
 */
export interface BufferView {
  /**
   * The index of the buffer.
   */
  buffer: number
  /**
   * The total byte length of the buffer view.
   */
  byteLength: number
  /**
   * The offset into the buffer in bytes.
   */
  byteOffset?: number
  /**
   * The stride, in bytes.
   */
  byteStride?: number
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * The target that the GPU buffer should be bound to.
   */
  target?: number
}

/**
 * A buffer points to binary geometry, animation, or skins.
 */
export interface Buffer {
  /**
   * The length of the buffer in bytes.
   */
  byteLength: number
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * The uri of the buffer.
   */
  uri?: string
}

/**
 * A camera's projection.  A node can reference a camera to apply a transform to place the
 * camera in the scene.
 */
export interface Camera {
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * An orthographic camera containing properties to create an orthographic projection matrix.
   */
  orthographic?: CameraOrthographic
  /**
   * A perspective camera containing properties to create a perspective projection matrix.
   */
  perspective?: CameraPerspective
  /**
   * Specifies if the camera uses a perspective or orthographic projection.
   */
  type: string
}
// ...
export interface Image {
  /**
   * The index of the bufferView that contains the image. Use this instead of the image's uri
   * property.
   */
  bufferView?: number
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The image's MIME type. Required if `bufferView` is defined.
   */
  mimeType?: string
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * The uri of the image.
   */
  uri?: string
}

/**
 * The material appearance of a primitive.
 */
export interface Material {
  /**
   * The alpha cutoff value of the material.
   */
  alphaCutoff?: number
  /**
   * The alpha rendering mode of the material.
   */
  alphaMode?: string
  /**
   * Specifies whether the material is double sided.
   */
  doubleSided?: boolean
  /**
   * The emissive color of the material.
   */
  emissiveFactor?: number[]
  /**
   * The emissive map texture.
   */
  emissiveTexture?: PurpleTextureInfo
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * The normal map texture.
   */
  normalTexture?: MaterialNormalTextureInfoObject
  /**
   * The occlusion map texture.
   */
  occlusionTexture?: MaterialOcclusionTextureInfoObject
  /**
   * A set of parameter values that are used to define the metallic-roughness material model
   * from Physically-Based Rendering (PBR) methodology. When not specified, all the default
   * values of `pbrMetallicRoughness` apply.
   */
  pbrMetallicRoughness?: MaterialPBRMetallicRoughness
}
// ...
export interface Mesh {
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * An array of primitives, each defining geometry to be rendered with a material.
   */
  primitives: MeshPrimitive[]
  /**
   * Array of weights to be applied to the Morph Targets.
   */
  weights?: number[]
}
// ...
export interface Node {
  /**
   * The index of the camera referenced by this node.
   */
  camera?: number
  /**
   * The indices of this node's children.
   */
  children?: number[]
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * A floating-point 4x4 transformation matrix stored in column-major order.
   */
  matrix?: number[]
  /**
   * The index of the mesh in this node.
   */
  mesh?: number
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * The node's unit quaternion rotation in the order (x, y, z, w), where w is the scalar.
   */
  rotation?: number[]
  /**
   * The node's non-uniform scale, given as the scaling factors along the x, y, and z axes.
   */
  scale?: number[]
  /**
   * The index of the skin referenced by this node.
   */
  skin?: number
  /**
   * The node's translation along the x, y, and z axes.
   */
  translation?: number[]
  /**
   * The weights of the instantiated Morph Target. Number of elements must match number of
   * Morph Targets of used mesh.
   */
  weights?: number[]
}

/**
 * Texture sampler properties for filtering and wrapping modes.
 */
export interface Sampler {
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * Magnification filter.
   */
  magFilter?: number
  /**
   * Minification filter.
   */
  minFilter?: number
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * s wrapping mode.
   */
  wrapS?: number
  /**
   * t wrapping mode.
   */
  wrapT?: number
}

/**
 * The root nodes of a scene.
 */
export interface Scene {
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * The indices of each root node.
   */
  nodes?: number[]
}

/**
 * Joints and matrices defining a skin.
 */
export interface Skin {
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The index of the accessor containing the floating-point 4x4 inverse-bind matrices.  The
   * default is that each matrix is a 4x4 identity matrix, which implies that inverse-bind
   * matrices were pre-applied.
   */
  inverseBindMatrices?: number
  /**
   * Indices of skeleton nodes, used as joints in this skin.
   */
  joints: number[]
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * The index of the node used as a skeleton root.
   */
  skeleton?: number
}

/**
 * A texture and its sampler.
 */
export interface Texture {
  /**
   * Dictionary object with extension-specific objects.
   */
  extensions?: { [key: string]: { [key: string]: any } }
  /**
   * Application-specific data.
   */
  extras?: any
  /**
   * The user-defined name of this object.
   */
  name?: string
  /**
   * The index of the sampler used by this texture. When undefined, a sampler with repeat
   * wrapping and auto filtering should be used.
   */
  sampler?: number
  /**
   * The index of the image used by this texture. When undefined, it is expected that an
   * extension or other mechanism will supply an alternate texture source, otherwise behavior
   * is undefined.
   */
  source?: number
}